divides code into modules
